requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
safetensors>=0.3.1
segment-anything>=1.0
pyyaml>=6.0
tqdm>=4.65.0
pillow>=9.0.0
einops>=0.6.1 